requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pinecone>=7.0.0
google-generativeai>=0.8.0
PyPDF2==3.0.1
python-docx==1.1.0
sentence-transformers>=3.0.0
numpy>=1.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2
tiktoken>=0.8.0